[Productionisation] Multi Executor Prototype with offline data nodes
Problem
=======
1. All sensor nodes are to be created using a single sensors executable.
ROS2 APIs have to be leveraged to build the same.
2. Furthermore, offline data processing capabilities are missing

Solution
========
ROS2 provides a `MultiThreadedExecutor` which can be used to control
the execution of callbacks of multiple nodes. However, publishers as
implemented in the sensor lib are not compatible with the aforementioned
API. Thus, multi-threaded prototypes for offline data publishers have
been implemented and spawned using the `MultiThreadedExecutor` API
- Added `SimpleImagePublisher` class
- Added `VideoDataReader` class
- Added `DayGpsFileReader` class
- Implemented `MultiThreadedExecutor` in `publisher_member_function.cpp`
- Updated Node name to "talker" in `sensor_launch.py`
- Added Header Guards in header files to prevent redefinition errors

Note
====
The real advantage of the `MultiThreadedExecutor` will be seen when implementing
subscriber groups.